chore: move ggez_utility to pic module
